+++ CHANGELOG.md
@@ -31,6 +31,7 @@
 * Rename `EveOnline::ESI::UniverseRegions#universe_regions_ids` to `#universe_region_ids`
 * Rename `EveOnline::ESI::UniverseConstellations#constellations_ids` to `#constellation_ids`
 * Rename `EveOnline::ESI::DogmaAttributes#attributes` to `#attribute_ids`
+* Rename `EveOnline::ESI::CharacterImplants#implants` to `#implant_ids`
 
 ## v0.18.0